requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
requests
jinja2
